feat : add api swagger decorator for authentication and guards on controllers and improve the implementation of all modules

+++ app/src/user/user.controller.ts
@@ -8,6 +8,9 @@ import {
   HttpCode,
   HttpStatus,
   SerializeOptions,
+  Request,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dtos/user-input.dto';
@@ -15,6 +18,8 @@ import { ApiBearerAuth, ApiOperation, ApiResponse } from '@nestjs/swagger';
 import { UserResponseDto } from './dtos/user-response.dto';
 import { CatResponseDto } from '@/cat/dtos';
 import { CatService } from '@/cat/cat.service';
+import RequestUserInformations from '@/interfaces/request-user-informations.interface';
+import { AuthGuard } from '@nestjs/passport';
 
 @Controller('user') // route '/user'
 export class UserController {
@@ -25,7 +30,6 @@ export class UserController {
 
   @Get() // GET '/user'
   @SerializeOptions({ type: UserResponseDto })
-  @ApiBearerAuth()
   @ApiOperation({ summary: 'Get all users' })
   @ApiResponse({
     status: 200,
@@ -63,6 +67,8 @@ export class UserController {
   }
 
   @Patch(':id') // PATCH '/user/:id'
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
   @SerializeOptions({ type: UserResponseDto })
   @ApiOperation({ summary: 'Update a user' })
   @ApiResponse({
@@ -73,15 +79,31 @@ export class UserController {
   update(
     @Param('id') id: string,
     @Body() updateUserDto: UpdateUserDto,
+    @Request() req: RequestUserInformations,
   ): Promise<UserResponseDto> {
+    if (req.user.userId !== id) {
+      throw new UnauthorizedException(
+        'You cannot update the information of a user other than yourself',
+      );
+    }
     return this.userService.update(id, updateUserDto);
   }
 
   @Delete(':id') // DELETE '/user/:id'
+  @UseGuards(AuthGuard('jwt'))
+  @ApiBearerAuth()
   @HttpCode(HttpStatus.NO_CONTENT)
   @ApiOperation({ summary: 'Delete a user' })
   @ApiResponse({ status: HttpStatus.NO_CONTENT })
-  async remove(@Param('id') id: string): Promise<void> {
+  async remove(
+    @Param('id') id: string,
+    @Request() req: RequestUserInformations,
+  ): Promise<void> {
+    if (req.user.userId !== id) {
+      throw new UnauthorizedException(
+        'You cannot delete the information of a user other than yourself',
+      );
+    }
     await this.userService.remove(id);
   }
 }
